Conflicts:
	dune/stuff/grid/provider/default.hh

+++ dune/stuff/la/solver/istl_amg.hh
@@ -11,10 +11,10 @@
 
 #if HAVE_DUNE_ISTL
 #include <dune/istl/operators.hh>
-#include <dune/istl/preconditioners.hh>
 #include <dune/istl/solvers.hh>
 #include <dune/stuff/common/disable_warnings.hh>
 #include <dune/istl/paamg/amg.hh>
+#include <dune/istl/preconditioners.hh>
 #include <dune/stuff/common/reenable_warnings.hh>
 #endif // HAVE_DUNE_ISTL
 
@@ -133,14 +133,14 @@ public:
     typedef SeqILU0<IstlMatrixType, IstlVectorType, IstlVectorType> SmootherType;
 
     typename Amg::SmootherTraits<SmootherType>::Arguments smoother_parameters;
-    smoother_parameters.iterations = opts.get("smoother.iterations", default_opts.get<size_t>("smoother.iterations"));
+    smoother_parameters.iterations = opts.get("smoother.iterations", default_opts.get<int>("smoother.iterations"));
     smoother_parameters.relaxationFactor =
         opts.get("smoother.relaxation_factor", default_opts.get<S>("smoother.relaxation_factor"));
 
     // define the AMG as the preconditioner for the BiCGStab solver
     Amg::Parameters amg_parameters(
-        opts.get("preconditioner.max_level", default_opts.get<size_t>("preconditioner.max_level")),
-        opts.get("preconditioner.coarse_target", default_opts.get<size_t>("preconditioner.coarse_target")),
+        opts.get("preconditioner.max_level", default_opts.get<int>("preconditioner.max_level")),
+        opts.get("preconditioner.coarse_target", default_opts.get<int>("preconditioner.coarse_target")),
         opts.get("preconditioner.min_coarse_rate", default_opts.get<S>("preconditioner.min_coarse_rate")),
         opts.get("preconditioner.prolong_damp", default_opts.get<S>("preconditioner.prolong_damp")));
     amg_parameters.setDefaultValuesIsotropic(
@@ -158,7 +158,7 @@ public:
                                           scalar_product,
                                           preconditioner,
                                           opts.get("precision", default_opts.get<S>("precision")),
-                                          opts.get("max_iter", default_opts.get<size_t>("max_iter")),
+                                          opts.get("max_iter", default_opts.get<int>("max_iter")),
                                           opts.get("verbose", default_opts.get<int>("verbose")));
     InverseOperatorResult stats;
     solver.apply(solution.backend(), rhs.backend(), stats);
